backend/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Flask==2.2.2
Flask-Cors==3.0.10
gitdb==4.0.10
GitPython==3.1.30
greenlet>=2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy>=1.26.4
pandas>=2.2.1
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
typing_extensions==4.5.0
tzdata==2024.1
Werkzeug==2.2.2
python-Levenshtein
